Make URL reflect search keyword
